resolved int overflow bug for phone number and also added optimistic ui support

+++ backend/models/customer.py
@@ -1,5 +1,5 @@
 from core.db import Base
-from sqlalchemy import Integer, String
+from sqlalchemy import BigInteger, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 
@@ -9,4 +9,4 @@ class Customer(Base):
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     email: Mapped[str] = mapped_column(String, unique=True, nullable=False)
     full_name: Mapped[str] = mapped_column(String, nullable=False)
-    phone_no: Mapped[int] = mapped_column(Integer, nullable=False)
+    phone_no: Mapped[int] = mapped_column(BigInteger, nullable=False)
